Eliminate GHC.Internal.List_LHAssumptions
