+++ src/Data/List_LHAssumptions.hs
@@ -1,4 +1,9 @@
 {-# OPTIONS_GHC -fplugin=LiquidHaskellBoot #-}
 module Data.List_LHAssumptions where
 
-import GHC.Internal.List_LHAssumptions()
+-- TODO: For some reason, the specifications of GHC.List have
+-- a role when verifying functions from Data.List. e.g
+-- basic/pos/AssmReflFilter.hs
+--
+-- Needs to be investigated.
+import GHC.List_LHAssumptions()
